requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
djangorestframework==3.14.0
numpy==1.23.4
pandas==1.5.0
pytrends==4.8.0
